# conception/api_swagger.yaml
openapi: 3.0.3
info:
  title: Spark
  description: |-
    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
  termsOfService: ///
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /associations/{assoID}:
    put:
      tags:
        - associations
      summary: edit profile
      description: edit the profile to modify it
      operationId: editAssoProfile
      parameters:
        - name: assoID
          in: path
          description: ID
          required: true
          schema:
            type: integer
            format: ID
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid input  
    get:
      tags:
        - associations
      summary: get asso profile
      description: get the profile
      operationId: getAssoProfile
      parameters:
        - name: assoID
          in: path
          description: ID
          required: true
          schema:
            type: integer
            format: ID
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid input           
  /calendar:
    post:
      tags:
        - calendar
      summary: Add a new event to the calendar
      description: Add a new event to the calendar
      operationId: addEvent
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
...
  /calendar/event/findByTags:
    get:
      tags:
        - calendar
      summary: Finds Events by tags
      description: Using tags as "association event", "enterprise event", "personal event"
      operationId: findEventsByTags
      parameters:
        - name: tags
          in: query
          description: Using "association event", "enterprise event", "personal event"
          required: false
          explode: true
          schema:
            type: array
            items: 
              type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid tag value
  /calendar/{eventID}:
    get:
      tags:
        - calendar
      summary: Find event by ID
      description: Returns a single event
      operationId: getEventById
      parameters:
        - name: eventID
          in: path
          description: ID of event to return
          required: true
          schema:
            type: integer
            format: eventID
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Event not found
    delete:
      tags:
        - calendar
      summary: Deletes an event from the calendar
      description: delete an event from the calendar
      operationId: deleteEEvent
      parameters:
        - name: eventID
          in: path
          description: Event id to delete
          required: true
          schema:
            type: integer
            format: eventID
      responses:
        '400':
          description: Invalid input
...
  /events/{EventID}:
    get:
      tags:
        - events
      summary: Returns event by ID
      description: Returns event by ID
      operationId: getEvent
      parameters:
        - name: EventID
          in: path
          description: Event id to get
          required: true
          schema:
            type: integer
            format: eventID
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid input
    put:
      tags:
        - events
      summary: Edit event by ID
      description: Edit event by ID
      operationId: editEvent
      parameters:
        - name: EventID
          in: path
          description: Event id to modify
          required: true
          schema:
            type: object
            format: eventType
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid input
    delete:
      tags:
        - events
      summary: Delete an event by ID
      description: Delete an Event
      operationId: deleteEvent
      parameters:
        - name: EventID
          in: path
          description: Event id to delete
          required: true
          schema:
            type: integer
            format: eventID
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid input
...